Extract parse method to inside MvxColor
