+++ MvvmCross-Plugins/Color/MvvmCross.Plugins.Color/MvxARGBValueConverter.cs
@@ -11,13 +11,9 @@ using MvvmCross.Platform.UI;
 namespace MvvmCross.Plugins.Color
 {
     [Preserve(AllMembers = true)]
-	public class MvxARGBValueConverter : MvxRGBValueConverter
+    public class MvxARGBValueConverter : MvxColorValueConverter<string>
     {
-        protected override MvxColor Parse8DigitColor(string value)
-        {
-            var a = int.Parse(value.Substring(0, 2), NumberStyles.HexNumber);
-            var rgb = int.Parse(value.Substring(2, 6), NumberStyles.HexNumber);
-            return new MvxColor(rgb, a);
-        }
+        protected override MvxColor Convert(string value, object parameter, CultureInfo culture)
+            => MvxColor.ParseHexString(value, assumeArgb: true);
     }
 }
