+++ MvvmCross-Plugins/Color/MvvmCross.Plugins.Color/MvxRGBValueConverter.cs
@@ -14,58 +14,6 @@ namespace MvvmCross.Plugins.Color
 	public class MvxRGBValueConverter : MvxColorValueConverter<string>
     {
         protected override MvxColor Convert(string value, object parameter, CultureInfo culture)
-        {
-            if (string.IsNullOrEmpty(value))
-                return new MvxColor(0);
-
-            value = value.TrimStart('#');
-            if (value.Length == 0)
-                return new MvxColor(0);
-
-            switch (value.Length)
-            {
-                case 3:
-                    return Parse3DigitColor(value);
-
-                case 6:
-                    return Parse6DigitColor(value);
-
-                case 8:
-                    return Parse8DigitColor(value);
-
-                default:
-                    return new MvxColor(0);
-            }
-        }
-
-        private MvxColor Parse3DigitColor(string value)
-        {
-            var red = int.Parse(value.Substring(0, 1), NumberStyles.HexNumber);
-            var green = int.Parse(value.Substring(1, 1), NumberStyles.HexNumber);
-            var blue = int.Parse(value.Substring(2, 1), NumberStyles.HexNumber);
-            return new MvxColor(UpByte(red), UpByte(green), UpByte(blue));
-        }
-
-        private int UpByte(int input)
-        {
-            var fourBit = input & 0xF;
-            var output = fourBit << 4;
-            output |= fourBit;
-            return output;
-        }
-
-        private MvxColor Parse6DigitColor(string value)
-        {
-            var rgb = int.Parse(value, NumberStyles.HexNumber);
-            return new MvxColor(rgb, 255);
-        }
-
-        protected virtual MvxColor Parse8DigitColor(string value)
-        {
-            // assume RGBA
-            var rgb = int.Parse(value.Substring(0, 6), NumberStyles.HexNumber);
-            var a = int.Parse(value.Substring(6, 2), NumberStyles.HexNumber);
-            return new MvxColor(rgb, a);
-        }
+            => MvxColor.ParseHexString(value);
     }
 }
