+++ MvvmCross/Platform/Platform/UI/MvxColor.cs
@@ -5,6 +5,8 @@
 //
 // Project Lead - Stuart Lodge, @slodge, me@slodge.com
 
+using System.Globalization;
+
 namespace MvvmCross.Platform.UI
 {
     public class MvxColor
@@ -79,6 +81,91 @@ namespace MvvmCross.Platform.UI
         public override string ToString()
         {
             return $"argb: #{A:X2}{R:X2}{G:X2}{B:X2}";
+        }
+
+        public static MvxColor ParseHexString(string value)
+            => ParseHexString(value, assumeArgb: false);
+
+        public static MvxColor ParseHexString(string value, bool assumeArgb)
+        {
+            if (string.IsNullOrEmpty(value))
+                return new MvxColor(0);
+
+            value = value.TrimStart('#');
+            if (value.Length == 0)
+                return new MvxColor(0);
+
+            switch (value.Length)
+            {
+                case 3:
+                    return Parse3DigitColor(value);
+
+                case 4:
+                    return assumeArgb ? Parse4DigitARBGColor(value) : Parse4DigitRBGAColor(value);
+
+                case 6:
+                    return Parse6DigitColor(value);
+
+                case 8:
+                    return assumeArgb ? Parse8DigitARGBColor(value) : Parse8DigitRGBAColor(value);
+
+                default:
+                    return new MvxColor(0);
+            }
+        }
+
+        private static int UpByte(int input)
+        {
+            var fourBit = input & 0xF;
+            var output = fourBit << 4;
+            output |= fourBit;
+            return output;
+        }
+
+        private static MvxColor Parse3DigitColor(string value)
+        {
+            var red = int.Parse(value.Substring(0, 1), NumberStyles.HexNumber);
+            var green = int.Parse(value.Substring(1, 1), NumberStyles.HexNumber);
+            var blue = int.Parse(value.Substring(2, 1), NumberStyles.HexNumber);
+            return new MvxColor(UpByte(red), UpByte(green), UpByte(blue));
+        }
+
+        private static MvxColor Parse6DigitColor(string value)
+        {
+            var rgb = int.Parse(value, NumberStyles.HexNumber);
+            return new MvxColor(rgb, 255);
+        }
+
+        private static MvxColor Parse4DigitARBGColor(string value)
+        {
+            var alpha = int.Parse(value.Substring(0, 1), NumberStyles.HexNumber);
+            var red = int.Parse(value.Substring(1, 1), NumberStyles.HexNumber);
+            var green = int.Parse(value.Substring(2, 1), NumberStyles.HexNumber);
+            var blue = int.Parse(value.Substring(3, 1), NumberStyles.HexNumber);
+            return new MvxColor(UpByte(red), UpByte(green), UpByte(blue), UpByte(alpha));
+        }
+
+        private static MvxColor Parse4DigitRBGAColor(string value)
+        {
+            var red = int.Parse(value.Substring(0, 1), NumberStyles.HexNumber);
+            var green = int.Parse(value.Substring(1, 1), NumberStyles.HexNumber);
+            var blue = int.Parse(value.Substring(2, 1), NumberStyles.HexNumber);
+            var alpha = int.Parse(value.Substring(3, 1), NumberStyles.HexNumber);
+            return new MvxColor(UpByte(red), UpByte(green), UpByte(blue), UpByte(alpha));
+        }
+
+        private static MvxColor Parse8DigitARGBColor(string value)
+        {
+            var a = int.Parse(value.Substring(0, 2), NumberStyles.HexNumber);
+            var rgb = int.Parse(value.Substring(2, 6), NumberStyles.HexNumber);
+            return new MvxColor(rgb, a);
+        }
+
+        private static MvxColor Parse8DigitRGBAColor(string value)
+        {
+            var rgb = int.Parse(value.Substring(0, 6), NumberStyles.HexNumber);
+            var a = int.Parse(value.Substring(6, 2), NumberStyles.HexNumber);
+            return new MvxColor(rgb, a);
         }
     }
 }
